+++ src/components/Auth/Login/useLogin.js
@@ -17,7 +17,7 @@ const useLogin = () => {
   const navigate = useNavigate();
   // === api ===
   const { data, statusCode, fetchData } = useAxios(
-    "/api/login", // 로그인 api 주소, 추후 백엔드 서버 구축 후 연결
+    "/account/login", // 로그인 api 주소, 추후 백엔드 서버 구축 후 연결
     "POST",
     {},
     false,
@@ -57,6 +57,7 @@ const useLogin = () => {
     }
   };
 
+  // === return ===
   return { idRef, passwordRef, submitRef, loginError, handleLogin };
 };
